Authorization: Deleting, Update,GetById a User's Task

+++ src/tasks/tasks.service.ts
@@ -20,8 +20,8 @@ export class TasksService {
 		return this.taskRepository.getTasks(filterDto, user);
 	}
 
-	async getTaskById(id: number): Promise<Task> {
-		const found = await this.taskRepository.findOne(id);
+	async getTaskById(id: number, user: User): Promise<Task> {
+		const found = await this.taskRepository.findOne({ where: { user, userId: user.id } });
 
 		if (!found) {
 			throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -30,8 +30,8 @@ export class TasksService {
 		return found;
 	}
 
-	async deleteTask(id: number): Promise<DeleteResult> {
-		const deleteResult = await this.taskRepository.delete(id);
+	async deleteTask(id: number, user: User): Promise<DeleteResult> {
+		const deleteResult = await this.taskRepository.delete({ id, userId: user.id });
 
 		if (deleteResult.affected === 0) {
 			throw new NotFoundException(`Task with ID "${id}" not found`)
@@ -39,8 +39,8 @@ export class TasksService {
 		return deleteResult;
 	}
 
-	async updateTaskStatus(id: number, status: TaskStatus) {
-		const task = await this.getTaskById(id);
+	async updateTaskStatus(id: number, status: TaskStatus, user: User) {
+		const task = await this.getTaskById(id, user);
 		task.status = status;
 		task.save();
 		return task;
